Safari-609.1.11.1/LayoutTests/media/modern-media-controls/status-label/status-label-font-expected.txt

Testing the StatusLabel font is correctly set when hosted in media controls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(mediaControls.statusLabel.element).fontFamily is "-apple-system"

PASS successfullyParsed is true

TEST COMPLETE
